up_route_api/src/requirements.txt
Flask[async]==3.0.3
requests==2.32.3
ortools==9.10.4067
folium==0.17.0
gunicorn==23.0.0
psycopg[binary]==3.2.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1